Viewer MVVM phase 5: extract MapInteractionController
All map gesture, pointer, and viewport-readout wiring moves out of
MainWindow into a dedicated controller in Services/. The controller
owns:

* Trackpad gestures (magnify, rotate, wheel-pan).
* Single/double-tap, modifier-click, and long-press pick dispatch
  (with the long-press timer + move-tolerance state).
* Pointer pressed/moved/exited/released wiring including the
  KeyModifiers capture used by tap-pick.
* Mouse lat/lon readout (cursor tracking + viewport-change refresh).
* Scale-bar and compass-rose viewport sync.
* Zoom in/out overlay buttons.
* Compass-rose RotationRequested / RotationResetRequested.

MainWindow now configures the whole interaction surface in one
new MapInteractionController(_viewModel, _pickService, _loader)
    .Attach(MapControl, ZoomInButton, ZoomOutButton, ScaleBar, CompassRose)
call. The Esc → ExitPickMode shortcut, pick-mode cursor/button
styling, and pick-mode toolbar button click stay in the window
since they're pick-mode UI affordances rather than gesture wiring
(deferred to a possible later phase as an attached behavior).

MainWindow.axaml.cs: 811 → 421 lines (1200 → 421 across all phases).
52/52 viewer tests pass; full solution builds clean.
